latest

+++ src/scene/shader-lib/wgsl/chunks/gsplat/vert/gsplatCommon.js
@@ -14,8 +14,8 @@ export default /* wgsl */`
 
 // modify the gaussian corner so it excludes gaussian regions with alpha less than 1/255
 fn clipCorner(corner: ptr<function, SplatCorner>, alpha: f32) {
-    let clip: f32 = min(1.0, sqrt(-log(1.0 / (255.0 * alpha))) / 2.0);
-    corner.offset = corner.offset * clip;
+    let clip: half = half(min(1.0, sqrt(-log(1.0 / (255.0 * alpha))) / 2.0));
+    corner.offset = corner.offset * f32(clip);
     corner.uv = corner.uv * clip;
 }
 `;
